* 6.4:
  [DoctrineBridge] Fix global query time calculation in profiler

+++ src/Symfony/Bridge/Doctrine/DataCollector/DoctrineDataCollector.php
@@ -85,7 +85,7 @@ class DoctrineDataCollector extends DataCollector
         return $this->data['queries'];
     }
 
-    public function getTime(): int
+    public function getTime(): float
     {
         $time = 0;
         foreach ($this->data['queries'] as $queries) {
